+++ rosrust/examples/publisher.rs
@@ -10,6 +10,7 @@ fn main() {
 
     // Create publisher
     let chatter_pub = rosrust::publish("chatter", 2).unwrap();
+    chatter_pub.wait_for_subscribers(None).unwrap();
 
     let log_names = rosrust::param("~log_names").unwrap().get().unwrap_or(false);