+++ rosrust/tests/publisher_to_roscpp_subscriber.rs
@@ -27,6 +27,7 @@ fn publisher_to_roscpp_subscriber() {
         .unwrap();
 
     let publisher = rosrust::publish::<msg::std_msgs::String>("chatter", 100).unwrap();
+    publisher.wait_for_subscribers(None).unwrap();
 
     let message = msg::std_msgs::String {
         data: "hello world".into(),